Moves performance tests to be run on demand
